Finalizado TP1

+++ Cliente.cs
@@ -2,17 +2,17 @@ public class Cliente {
 
     string nombre;
     string direccion;
-    int numero;
+    double numero;
     string datosObsDirec;
 
-    public Cliente(string nombre, string direccion, int numero)
+    public Cliente(string nombre, string direccion, double numero)
     {
         this.nombre = nombre;
         this.direccion = direccion;
         this.numero = numero;
     }
 
-    public Cliente(string nombre, string direccion, int numero, string datosObsDirec)
+    public Cliente(string nombre, string direccion, double numero, string datosObsDirec)
     {
         this.nombre = nombre;
         this.direccion = direccion;
@@ -22,6 +22,6 @@ public class Cliente {
 
     public string Nombre { get => nombre; set => nombre = value; }
     public string Direccion { get => direccion; set => direccion = value; }
-    public int Numero { get => numero; set => numero = value; }
+    public double Numero { get => numero; set => numero = value; }
     public string DatosObsDirec { get => datosObsDirec; set => datosObsDirec = value; }
 }
